requirements.txt
fastapi
uvicorn
asyncpg
pydantic[email]
python-dotenv
passlib[bcrypt]
pyjwt